requirements.txt
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
packaging==24.2
pycparser==2.22
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
pandas==2.2.3
flask==3.1.0
flask-apscheduler
openpyxl
certifi==2024.12.14
charset-normalizer==3.4.1
crypto==1.4.1
idna==3.10
Naked==0.1.32
pycryptodome==3.21.0
PyYAML==6.0.2
requests==2.32.3
shellescape==3.8.1
urllib3==2.3.0
pyautogui>=0.9.0
pyautoit>=0.6.0